web-app/requirements.txt
fastapi==0.63.0
uvicorn==0.13.3 
aiofiles==0.6.0 
pydantic==1.7.3 
Jinja2==2.11.2 
pymongo==3.11.2 
dnspython==2.1.0 
python-multipart==0.0.5
grpcio==1.34.0 
grpcio-tools==1.34.0 